Formatting in code and output.
Unsure how we want key to be formatted for printing. Is "A3DF" printed as "xa3 xdf" or "0xA3,0xDF" or "A3DF"? More changes coming once this is settled.

Also, see line22 of 'ECDH and ECDSA Sign.py'. It reads [54:] but [64:] makes more sense intuitively. Similarly on line60.

I would change the way we are handling 'public_key_string' to get rid of the [:28], opting instead for a sandwich of "-beg-"->"(#key#)"->"-end-". Didn't change it because I'm not sure of things.

We can sit down or email to help me figure out some of these things.